moved GetBlockDeviceSelectorFields() to helper_blockdevice code
