+++ cmd/provisioner-localpv/app/config.go
@@ -25,7 +25,6 @@ import (
 	cast "github.com/openebs/maya/pkg/castemplate/v1alpha1"
 	hostpath "github.com/openebs/maya/pkg/hostpath/v1alpha1"
 	"github.com/openebs/maya/pkg/util"
-	"gopkg.in/yaml.v2"
 	klog "k8s.io/klog/v2"
 
 	//"github.com/pkg/errors"
@@ -140,9 +139,6 @@ const (
 	//KeyPVAbsolutePath = "AbsolutePath"
 )
 
-// BlockDeviceSelectorFields stores the block device selectors
-type BlockDeviceSelectorFields map[string]string
-
 const (
 	// Some of the PVCs launched with older helm charts, still
 	// refer to the StorageClass via beta annotations.
@@ -215,7 +211,7 @@ func (c *VolumeConfig) GetStorageType() string {
 	return stgType
 }
 
-//GetStorageType returns the BlockDeviceSelectors value configured
+//GetBlockDeviceSelectors returns the BlockDeviceSelectors value configured
 // in StorageClass. Default is ""
 func (c *VolumeConfig) GetBlockDeviceSelectors() string {
 	blockDeviceSelector := c.getValue(KeyBlockDeviceSelectors)
@@ -223,16 +219,6 @@ func (c *VolumeConfig) GetBlockDeviceSelectors() string {
 		return ""
 	}
 	return blockDeviceSelector
-}
-
-// GetBlockDeviceSelectorFields unmarshalls the multi-line blockDeviceSelectors string
-// value into a map of strings.
-func GetBlockDeviceSelectorFields(blockDeviceSelectors string) (map[string]string, error) {
-	var out BlockDeviceSelectorFields
-	if err := yaml.Unmarshal([]byte(blockDeviceSelectors), &out); err != nil {
-		return out, errors.Wrapf(err, "unable to unmarshal block device selector fields: {%s}", blockDeviceSelectors)
-	}
-	return out, nil
 }
 
 //GetFSType returns the FSType value configured
