+++ cmd/provisioner-localpv/app/helper_blockdevice.go
@@ -21,6 +21,8 @@ package app
 
 import (
 	"github.com/openebs/maya/pkg/util"
+	"gopkg.in/yaml.v2"
+
 	//"fmt"
 	//"path/filepath"
 	//"strings"
@@ -71,8 +73,11 @@ type HelperBlockDeviceOptions struct {
 
 	// bdSelectors stores the different fields
 	// used for selecting a block device
-	bdSelectors map[string]string
+	bdSelectors string
 }
+
+// BlockDeviceSelectorFields stores the block device selectors
+type BlockDeviceSelectorFields map[string]string
 
 // validate checks that the required fields to create BDC
 // are available
@@ -141,8 +146,13 @@ func (p *Provisioner) createBlockDeviceClaim(ctx context.Context, blkDevOpts *He
 	}
 
 	// if block device selectors are present, set it on the BDC
-	if len(blkDevOpts.bdSelectors) > 0 {
-		bdcObjBuilder.WithSelector(blkDevOpts.bdSelectors)
+	if blkDevOpts.bdSelectors != "" {
+		// convert the selector fields into a map
+		bdSelector, err := GetBlockDeviceSelectorFields(blkDevOpts.bdSelectors)
+		if err != nil {
+			return errors.Wrapf(err, "Failed to form BDC object")
+		}
+		bdcObjBuilder.WithSelector(bdSelector)
 	}
 
 	bdcObj, err := bdcObjBuilder.Build()
@@ -165,6 +175,16 @@ func (p *Provisioner) createBlockDeviceClaim(ctx context.Context, blkDevOpts *He
 	blkDevOpts.bdcName = bdcName
 
 	return nil
+}
+
+// GetBlockDeviceSelectorFields unmarshalls the multi-line blockDeviceSelectors string
+// value into a map of strings.
+func GetBlockDeviceSelectorFields(blockDeviceSelectors string) (map[string]string, error) {
+	var out BlockDeviceSelectorFields
+	if err := yaml.Unmarshal([]byte(blockDeviceSelectors), &out); err != nil {
+		return out, errors.Wrapf(err, "unable to unmarshal block device selector fields: {%s}", blockDeviceSelectors)
+	}
+	return out, nil
 }
 
 // getBlockDevicePath fetches the BDC associated with this Local PV
